fix:シェープファイルのパース時にfloatではなくdoubleで取得

+++ PlateauToolkit.Sandbox/Editor/PlateauSandboxPrefabPlacement.cs
@@ -20,8 +20,8 @@ namespace PlateauToolkit.Sandbox.Editor
 
         public struct PlacementContext
         {
-            public float m_Longitude;
-            public float m_Latitude;
+            public double m_Longitude;
+            public double m_Latitude;
             public float m_Height;
             public GameObject m_Prefab;
             public string m_AssetType;
